Removed all tRPC, prisma and supabase code. Deleted Auth session feature. Removed Image components. Removed env variables. Simplified Admin page
