+++ next.config.mjs
@@ -1,4 +1,3 @@
-import { env } from "./src/env/server.mjs";
 
 /**
  * Don't be scared of the generics here.
@@ -17,11 +16,11 @@ export default defineNextConfig({
   swcMinify: true,
   // Next.js i18n docs: https://nextjs.org/docs/advanced-features/i18n-routing
   i18n: {
-    locales: ["en"],
-    defaultLocale: "en",
+    locales: ["ru"],
+    defaultLocale: "ru",
   },
   images: {
-    domains: ['ntorzmsgnfxlxdfetvfz.supabase.co']
+    domains: []
   },
   typescript: {
     ignoreBuildErrors: true
