+++ package.json
@@ -12,28 +12,18 @@
   "dependencies": {
     "@mdx-js/loader": "^2.2.1",
     "@mdx-js/react": "^2.2.1",
-    "@next-auth/prisma-adapter": "^1.0.4",
     "@next/mdx": "^13.1.5",
-    "@prisma/client": "^4.3.1",
-    "@supabase/supabase-js": "^2.0.4",
-    "@trpc/client": "^9.27.2",
-    "@trpc/next": "^9.27.2",
-    "@trpc/react": "^9.27.2",
-    "@trpc/server": "^9.27.2",
     "@vercel/analytics": "^0.1.6",
     "dayjs": "^1.11.7",
     "gray-matter": "^4.0.3",
     "next": "^13.1.6",
-    "next-auth": "^4.18.4",
     "next-mdx-remote": "^4.3.0",
     "next-seo": "^5.15.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-query": "^3.39.2",
     "reading-time": "^1.5.0",
-    "rehype-autolink-headings": "^6.1.1",
-    "superjson": "^1.9.1",
-    "zod": "^3.18.0"
+    "rehype-autolink-headings": "^6.1.1"
   },
   "devDependencies": {
     "@types/node": "18.0.0",
@@ -45,7 +35,6 @@
     "eslint": "8.22.0",
     "eslint-config-next": "^13.1.6",
     "postcss": "^8.4.16",
-    "prisma": "^4.3.1",
     "rehype-code-titles": "^1.1.0",
     "rehype-prism-plus": "^1.5.0",
     "rehype-slug": "^5.1.0",
