--- prisma/schema.prisma
@@ -1,39 +1,0 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
-generator client {
-    provider = "prisma-client-js"
-}
-
-datasource db {
-    provider = "postgresql"
-    // NOTE: When using postgresql, mysql or sqlserver, uncomment the @db.text annotations in model Account below
-    // Further reading: 
-    // https://next-auth.js.org/adapters/prisma#create-the-prisma-schema
-    // https://www.prisma.io/docs/reference/api-reference/prisma-schema-reference#string
-    url      = env("DATABASE_URL")
-}
-
-model Post {
-    id          Int       @id @default(autoincrement())
-    title       String
-    desc        String?
-    content     String?
-    published   Boolean   @default(false)
-    publishedAt DateTime?
-    createdAt   DateTime  @default(now())
-    image       String?
-}
-
-model User {
-    id    Int     @id @default(autoincrement())
-    email String  @unique
-    name  String?
-}
-
-model Photo {
-    id    Int     @id @default(autoincrement())
-    title String
-    desc  String?
-    src   String
-}
