+++ src/components/index.ts
@@ -3,10 +3,7 @@ export { default as Footer } from "./Footer";
 export { default as AboutItem } from "./AboutItem";
 export { default as Contacts } from "./Contacts";
 export { default as Layout } from "./Layout";
-export { default as AdminLayout } from "./AdminLayout";
-export { default as ImageWrapper } from "./ImageWrapper";
 export { default as Post } from "./Post";
-export { default as GalleryImage } from "./GalleryImage";
 export { default as MDXComponents } from "./MDXComponents";
 export { default as BlogHorizontalImage } from "./BlogHorizontalImage";
 export { default as BlogVerticalImage } from "./BlogVerticalImage";
