+++ src/pages/_app.tsx
@@ -1,85 +1,22 @@
-import { httpBatchLink } from "@trpc/client/links/httpBatchLink";
-import { loggerLink } from "@trpc/client/links/loggerLink";
-import { withTRPC } from "@trpc/next";
-import { SessionProvider } from "next-auth/react";
-import superjson from "superjson";
-import { AdminLayout, Layout, MDXComponents } from "../components";
+import { Layout, MDXComponents } from "../components";
 import { Analytics } from "@vercel/analytics/react";
 import { MDXProvider } from "@mdx-js/react";
 import type { AppType } from "next/dist/shared/lib/utils";
-import type { AppRouter } from "../server/router";
 import "../styles/globals.css";
 import "../styles/prism.css";
 
 const MyApp: AppType = ({
   Component,
-  pageProps: { session, ...pageProps },
-  router,
+  pageProps: { ...pageProps }
 }) => {
   return (
     <MDXProvider components={MDXComponents}>
-      <SessionProvider session={session}>
-        {router.pathname.includes("admin") ? (
-          <AdminLayout>
-            <Component {...pageProps} />
-          </AdminLayout>
-        ) : (
-          <Layout>
-            <Component {...pageProps} />
-          </Layout>
-        )}
-        <Analytics />
-      </SessionProvider>
+      <Layout>
+        <Component {...pageProps} />
+      </Layout>
+      <Analytics />
     </MDXProvider>
   );
 };
 
-const getBaseUrl = () => {
-  if (typeof window !== "undefined") return ""; // browser should use relative url
-  if (process.env.VERCEL_URL) return `https://${process.env.VERCEL_URL}`; // SSR should use vercel url
-  return `http://localhost:${process.env.PORT ?? 3000}`; // dev SSR should use localhost
-};
-
-export default withTRPC<AppRouter>({
-  config({ ctx }) {
-    /**
-     * If you want to use SSR, you need to use the server's full URL
-     * @link https://trpc.io/docs/ssr
-     */
-    const url = `${getBaseUrl()}/api/trpc`;
-
-    return {
-      links: [
-        loggerLink({
-          enabled: (opts) =>
-            process.env.NODE_ENV === "development" ||
-            (opts.direction === "down" && opts.result instanceof Error),
-        }),
-        httpBatchLink({ url }),
-      ],
-      url,
-      transformer: superjson,
-      /**
-       * @link https://react-query.tanstack.com/reference/QueryClient
-       */
-      // queryClientConfig: { defaultOptions: { queries: { staleTime: 60 } } },
-
-      // To use SSR properly you need to forward the client's headers to the server
-      // headers: () => {
-      //   if (ctx?.req) {
-      //     const headers = ctx?.req?.headers;
-      //     delete headers?.connection;
-      //     return {
-      //       ...headers,
-      //       "x-ssr": "1",
-      //     };
-      //   }
-      //   return {};
-      // }
-    };
-  },
-  /**
-   * @link https://trpc.io/docs/ssr
-   */
-  ssr: false,
-})(MyApp);
+export default MyApp
