--- src/server/router/example.ts
@@ -1,21 +1,0 @@
-import { createRouter } from "./context";
-import { z } from "zod";
-
-export const exampleRouter = createRouter()
-  .query("hello", {
-    input: z
-      .object({
-        text: z.string().nullish(),
-      })
-      .nullish(),
-    resolve({ input }) {
-      return {
-        greeting: `Hello ${input?.text ?? "world"}`,
-      };
-    },
-  })
-  .query("getAll", {
-    async resolve({ ctx }) {
-      return await ctx.prisma?.photo.findMany();
-    },
-  });
