--- src/server/router/protected-example-router.ts
@@ -1,14 +1,0 @@
-import { createProtectedRouter } from "./context";
-
-// Example router with queries that can only be hit if the user requesting is signed in
-export const protectedExampleRouter = createProtectedRouter()
-  .query("getSession", {
-    resolve({ ctx }) {
-      return ctx.session;
-    },
-  })
-  .query("getSecretMessage", {
-    resolve({ ctx }) {
-      return "He who asks a question is a fool for five minutes; he who does not ask a question remains a fool forever.";
-    },
-  });
