--- src/supabase/supabaseClient.ts
@@ -1,7 +1,0 @@
-import { createClient } from '@supabase/supabase-js'
-import { env } from "../env/client.mjs";
-
-const supabaseUrl = env.NEXT_PUBLIC_SUPABASE_URL || ''
-const supabaseAnonKey = env.NEXT_PUBLIC_SUPABASE_ANON_KEY || ''
-
-export const supabase = createClient(supabaseUrl, supabaseAnonKey)
